fix: scope unix lock descriptor test
