+++ internal/statelock/statelock_test.go
@@ -122,19 +122,3 @@ func TestOwnerMetadataRejectsSymlink(t *testing.T) {
 		t.Fatalf("victim metadata = %q, want %q", payload, want)
 	}
 }
-
-func TestOwnerMetadataFileRejectsNegativeDescriptor(t *testing.T) {
-	file, err := ownerMetadataFileFromDescriptor(-1, "owner.json")
-	if err == nil {
-		if file != nil {
-			_ = file.Close()
-		}
-		t.Fatal("ownerMetadataFileFromDescriptor() error = nil, want invalid descriptor rejection")
-	}
-	if file != nil {
-		t.Fatalf("ownerMetadataFileFromDescriptor() file = %v, want nil", file)
-	}
-	if !strings.Contains(err.Error(), "invalid descriptor") {
-		t.Fatalf("ownerMetadataFileFromDescriptor() error = %v, want invalid descriptor", err)
-	}
-}
